requirements.txt
flask==3.0.3
flask-swagger-ui==4.11.1
websocket-client==1.8.0
supabase==2.11.0
python-dotenv==1.0.0